refactor(runtime): extract retry module from agent_loop
The LLM-call retry path moves into `agent_loop/retry.rs`. Two
`pub(super)` entry points consumed by the main loop bodies:

- `call_with_retry` — non-streaming wrapper. Honors the
  `ProviderCooldown` circuit-breaker (Allow / AllowProbe / Reject),
  acquires `LLM_CONCURRENCY` per attempt so backoff sleeps don't pin
  a slot, classifies errors via `llm_errors` for sanitized
  user-facing messages, and appends the
  `librefang_channels::message_journal::RATE_LIMIT_DEFER_MARKER` to
  exhausted-retry errors so the channel bridge can route them to
  `JournalStatus::Deferred` instead of `Failed`.
- `stream_with_retry` — streaming variant returning
  `StreamWithRetryResult { response, cascade_leak_aborted }`. Same
  retry policy as the non-streaming path, plus the incremental
  cascade-leak guard that stops forwarding `TextDelta` events
  mid-stream when the in-memory accumulator trips
  `is_cascade_leak`.

Cluster-private helpers stay private to the module:
`check_retry_cooldown`, `record_retry_success`, `record_retry_failure`,
`handle_retryable_llm_error`, `build_user_facing_llm_error`.

Constants travel with their consumers: `MAX_RETRIES` (pub(super)),
`BASE_RETRY_DELAY_MS` (pub(super)), `MAX_CONCURRENT_LLM_CALLS`
(private), `DEFAULT_DEFER_MS` (private), and the `LLM_CONCURRENCY`
static. The two `pub(super)` constants are referenced from the
`mod tests` block and imported there via the existing
test-only `use super::retry::{...}` pattern.

`StreamWithRetryResult.response` and
`StreamWithRetryResult.cascade_leak_aborted` are `pub(super)` because
the streaming main loop in mod.rs destructures them after the call.

mod.rs sheds the now-orphan top-level imports that only the retry
path needed: `CooldownVerdict`, `ProviderCooldown`, `LlmError`, and
the entire `crate::llm_errors` use. retry.rs picks them up locally
along with `StopReason`, `TokenUsage`, and `super::TIMEOUT_PARTIAL_OUTPUT_MARKER`.

Behavior-preserving. mod.rs: 11,812 -> 11,384 LOC (-428).
cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
